+++ CHANGELOG.rst
@@ -48,6 +48,8 @@ TODO: integration test with pandapower (see `pandapower/contingency/contingency.
   `loader_kwargs` of LightSimBackend when loading it with pypowsybl)
 - [IMPROVED] Eigen to version 5.0.1 (2025/11/11)
 - [IMPROVED] rename all ".h" file to ".hpp" for cpp headers (cpp side). 
+- [IMPROVED] consistency of "bus labelling" cpp side (make it harder to 
+  mix `LocalBusId`, `GlobalBusId` / `GridModelBusId` and `SolverBusId`)
 
 [0.11.0] 2025-12-09
 ----------------------